backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yt-dlp==2024.8.6
mlx-whisper==0.4.2
python-docx==1.1.2
weasyprint==62.3
httpx==0.27.2
python-multipart==0.0.9
aiofiles==24.1.0
